chore: 식별자 타입 Integer -> Long으로 수정

+++ BE-MyCarMaster/src/main/java/softeer/bemycarmaster/api/domain/engine/controller/EngineController.java
@@ -27,7 +27,7 @@ public class EngineController {
 	@Operation(summary = "트림에 따른 엔진 목록을 반환합니다")
 	public Response<GetEnginesResponse> getEngines(@RequestBody @Valid GetEnginesRequest getEnginesRequest) {
 
-		Integer trimId = getEnginesRequest.getTrimId();
+		Long trimId = getEnginesRequest.getTrimId();
 		GetEnginesResponse getEnginesResponse = getEnginesUseCase.execute(trimId);
 		return Response.createSuccessResponse(getEnginesResponse);
 	}
